authblend/api/urls.py
"""URL configuration for API App."""

from django.urls import include, path
from rest_framework.routers import DefaultRouter
from api.views import UserViewSet

router = DefaultRouter()
router.register("users", UserViewSet)

urlpatterns = [
    path("", include(router.urls)),
]
