pyproject.toml
[tool.poetry]
name = "authblend"
version = "0.1.0"
description = ""
authors = ["umairkarel <umair.karel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
pre-commit = "^3.4.0"
black = "^23.9.1"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
